Refactor contracts tests to not use injected web3 instance
