+++ packages/0x.js/test/assert_test.ts
@@ -1,3 +1,4 @@
+import { web3Factory } from '@0xproject/dev-utils';
 import * as chai from 'chai';
 import 'mocha';
 
@@ -5,7 +6,6 @@ import { ZeroEx } from '../src';
 import { assert } from '../src/utils/assert';
 
 import { constants } from './utils/constants';
-import { web3Factory } from './utils/web3_factory';
 
 const expect = chai.expect;
