+++ packages/0x.js/test/utils/constants.ts
@@ -1,11 +1,9 @@
 export const constants = {
     NULL_ADDRESS: '0x0000000000000000000000000000000000000000',
-    RPC_URL: 'http://localhost:8545',
     ROPSTEN_NETWORK_ID: 3,
     KOVAN_NETWORK_ID: 42,
     TESTRPC_NETWORK_ID: 50,
     KOVAN_RPC_URL: 'https://kovan.infura.io/',
     ROPSTEN_RPC_URL: 'https://ropsten.infura.io/',
     ZRX_DECIMALS: 18,
-    GAS_ESTIMATE: 500000,
 };
