+++ packages/contracts/test/multi_sig_with_time_lock_except_remove_auth_addr.ts
@@ -1,7 +1,6 @@
-import { BlockchainLifecycle } from '@0xproject/dev-utils';
+import { BlockchainLifecycle, devConstants, web3Factory } from '@0xproject/dev-utils';
 import { Web3Wrapper } from '@0xproject/web3-wrapper';
 import * as chai from 'chai';
-import * as Web3 from 'web3';
 
 import * as tokenTransferProxyJSON from '../../build/contracts/TokenTransferProxy.json';
 import { Artifacts } from '../util/artifacts';
@@ -16,10 +15,8 @@ const PROXY_ABI = (tokenTransferProxyJSON as any).abi;
 
 chaiSetup.configure();
 const expect = chai.expect;
-// In order to benefit from type-safety, we re-assign the global web3 instance injected by Truffle
-// with type `any` to a variable of type `Web3`.
-const web3: Web3 = (global as any).web3;
-const blockchainLifecycle = new BlockchainLifecycle(constants.RPC_URL);
+const web3 = web3Factory.create();
+const blockchainLifecycle = new BlockchainLifecycle(devConstants.RPC_URL);
 
 describe('MultiSigWalletWithTimeLockExceptRemoveAuthorizedAddress', () => {
     const web3Wrapper = new Web3Wrapper(web3.currentProvider);
