+++ packages/contracts/util/constants.ts
@@ -3,7 +3,6 @@ export const constants = {
     INVALID_OPCODE: 'invalid opcode',
     REVERT: 'revert',
     TESTRPC_NETWORK_ID: 50,
-    RPC_URL: 'http://localhost:8545',
     MAX_ETHERTOKEN_WITHDRAW_GAS: 43000,
     MAX_TOKEN_TRANSFERFROM_GAS: 80000,
     MAX_TOKEN_APPROVE_GAS: 60000,
