+++ packages/dev-utils/src/constants.ts
@@ -1,0 +1,4 @@
+export const constants = {
+    RPC_URL: 'http://localhost:8545',
+    GAS_ESTIMATE: 500000,
+};
