+++ packages/dev-utils/src/globals.d.ts
@@ -1,0 +1,2 @@
+declare module 'web3-provider-engine';
+declare module 'web3-provider-engine/subproviders/rpc';
