+++ packages/dev-utils/src/index.ts
@@ -1,2 +1,4 @@
 export { RPC } from './rpc';
 export { BlockchainLifecycle } from './blockchain_lifecycle';
+export { web3Factory } from './web3_factory';
+export { constants as devConstants } from './constants';
