+++ packages/dev-utils/tsconfig.json
@@ -6,6 +6,7 @@
     "include": [
         "./src/**/*",
         "../../node_modules/types-bn/index.d.ts",
+        "../../node_modules/web3-typescript-typings/index.d.ts",
         "../../node_modules/types-ethereumjs-util/index.d.ts"
     ]
 }
